+++ frontend/app/src/main/java/com/ssafy/stellargram/ui/screen/chat/TestValue.kt
@@ -1,5 +1,5 @@
 package com.ssafy.stellargram.ui.screen.chat
 
 object TestValue {
-    val myId:Int = 111
+    val myId:Long = 111
 }
